Fix for adding app secret
As of this morning, it appears that the Azure CLI has changed such that we can no longer add an app secret the way we were previously doing it.

The fix is to make a separate az rest call to add the app secret. This approach is actually superior to the previous approach since the password is generated by Azure and it better encrypted than the one we were generating in Node